+++ apps/news/models/__init__.py
@@ -1,11 +1,9 @@
 from .observer_species_latest import ObserverSpeciesLatest
 from .observer_species_monthly import ObserverSpeciesMonthly
 from .observer_species_weekly import ObserverSpeciesWeekly
-from .year_list import YearList
 
 __all__ = [
     "ObserverSpeciesLatest",
     "ObserverSpeciesMonthly",
     "ObserverSpeciesWeekly",
-    "YearList",
 ]