+++ apps/species/apps.py
@@ -3,6 +3,5 @@ from django.utils.translation import gettext_lazy as _
 
 
 class Config(AppConfig):
-    label = "_species"
     name = "species"
     verbose_name = _("species.plural")